back_end/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
pandas==2.3.1
openpyxl==3.1.5
langchain_google_genai==2.1.9
langchain_experimental==0.3.4
tabulate==0.9.0
python-dotenv==1.1.1
pydantic==2.8.2

